Move a monitor_indexes::Db::get_index_version into a db actor
This is a part of #3.

The change moves get_index_version method from monitor_indexes into a db
actor. Next paches will move other methods into a db actor.
